feat: enable componentized global wrappers
